securing the homescreen through authentication
